+++ lib/page/mytests/theme_study/extension/myapp_theme_study.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:logger/logger.dart';
-import 'package:riverpod_modifiers_example/page/mytests/theme_study/homepage_theme_study.dart';
-import 'package:riverpod_modifiers_example/page/mytests/theme_study/theme_provider.dart';
+import 'package:riverpod_modifiers_example/page/mytests/theme_study/extension/homepage_theme_study.dart';
+import 'package:riverpod_modifiers_example/page/mytests/theme_study/extension/theme_provider.dart';
 
 final myLogger = Logger();
